+++ src/gpio.rs
@@ -3,7 +3,7 @@
 use core::convert::Infallible;
 use core::marker::PhantomData;
 
-use embedded_hal::digital::v2::{toggleable, InputPin, OutputPin, StatefulOutputPin};
+use embedded_hal::digital::v2::{InputPin, OutputPin, StatefulOutputPin, ToggleableOutputPin};
 
 use crate::pac::EXTI;
 use crate::syscfg::SysCfg;
@@ -81,7 +81,17 @@ pub struct PushPull;
 /// Analog mode (type state)
 pub struct Analog;
 
+/// Digital output pin state
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub enum PinState {
+    /// Low pin state
+    Low,
+    /// High pin state
+    High,
+}
+
 /// GPIO Pin speed selection
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum Speed {
     Low = 0,
     Medium = 1,
@@ -89,7 +99,7 @@ pub enum Speed {
     VeryHigh = 3,
 }
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum Edge {
     RISING,
     FALLING,
@@ -223,62 +233,150 @@ impl<MODE, const P: char> PinExt for PXx<MODE, P> {
     }
 }
 
-impl<MODE, const P: char> OutputPin for PXx<Output<MODE>, P> {
-    type Error = Infallible;
-
-    fn set_high(&mut self) -> Result<(), Self::Error> {
+impl<MODE, const P: char> PXx<Output<MODE>, P> {
+    #[inline(always)]
+    pub fn set_high(&mut self) {
         // NOTE(unsafe) atomic write to a stateless register
-        unsafe { (*Gpio::<P>::ptr()).bsrr.write(|w| w.bits(1 << self.i)) };
-        Ok(())
+        unsafe { (*Gpio::<P>::ptr()).bsrr.write(|w| w.bits(1 << self.i)) }
     }
 
-    fn set_low(&mut self) -> Result<(), Self::Error> {
+    #[inline(always)]
+    pub fn set_low(&mut self) {
         // NOTE(unsafe) atomic write to a stateless register
         unsafe {
             (*Gpio::<P>::ptr())
                 .bsrr
                 .write(|w| w.bits(1 << (self.i + 16)))
-        };
+        }
+    }
+
+    #[inline(always)]
+    pub fn get_state(&self) -> PinState {
+        if self.is_set_low() {
+            PinState::Low
+        } else {
+            PinState::High
+        }
+    }
+
+    #[inline(always)]
+    pub fn set_state(&mut self, state: PinState) {
+        match state {
+            PinState::Low => self.set_low(),
+            PinState::High => self.set_high(),
+        }
+    }
+
+    #[inline(always)]
+    pub fn is_set_high(&self) -> bool {
+        !self.is_set_low()
+    }
+
+    #[inline(always)]
+    pub fn is_set_low(&self) -> bool {
+        // NOTE(unsafe) atomic read with no side effects
+        unsafe { (*Gpio::<P>::ptr()).odr.read().bits() & (1 << self.i) == 0 }
+    }
+
+    #[inline(always)]
+    pub fn toggle(&mut self) {
+        if self.is_set_low() {
+            self.set_high()
+        } else {
+            self.set_low()
+        }
+    }
+}
+
+impl<MODE, const P: char> OutputPin for PXx<Output<MODE>, P> {
+    type Error = Infallible;
+
+    #[inline(always)]
+    fn set_high(&mut self) -> Result<(), Self::Error> {
+        self.set_high();
+        Ok(())
+    }
+
+    #[inline(always)]
+    fn set_low(&mut self) -> Result<(), Self::Error> {
+        self.set_low();
         Ok(())
     }
 }
 
 impl<MODE, const P: char> StatefulOutputPin for PXx<Output<MODE>, P> {
+    #[inline(always)]
     fn is_set_high(&self) -> Result<bool, Self::Error> {
-        self.is_set_low().map(|v| !v)
+        Ok(self.is_set_high())
     }
 
+    #[inline(always)]
     fn is_set_low(&self) -> Result<bool, Self::Error> {
-        // NOTE(unsafe) atomic read with no side effects
-        Ok(unsafe { (*Gpio::<P>::ptr()).odr.read().bits() & (1 << self.i) == 0 })
+        Ok(self.is_set_low())
     }
 }
 
-impl<MODE, const P: char> toggleable::Default for PXx<Output<MODE>, P> {}
+impl<MODE, const P: char> ToggleableOutputPin for PXx<Output<MODE>, P> {
+    type Error = Infallible;
+
+    #[inline(always)]
+    fn toggle(&mut self) -> Result<(), Self::Error> {
+        self.toggle();
+        Ok(())
+    }
+}
+
+impl<const P: char> PXx<Output<OpenDrain>, P> {
+    #[inline(always)]
+    fn is_high(&self) -> bool {
+        !self.is_low()
+    }
+
+    #[inline(always)]
+    fn is_low(&self) -> bool {
+        // NOTE(unsafe) atomic read with no side effects
+        unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << self.i) == 0 }
+    }
+}
 
 impl<const P: char> InputPin for PXx<Output<OpenDrain>, P> {
     type Error = Infallible;
 
+    #[inline(always)]
     fn is_high(&self) -> Result<bool, Self::Error> {
-        self.is_low().map(|v| !v)
+        Ok(self.is_high())
     }
 
+    #[inline(always)]
     fn is_low(&self) -> Result<bool, Self::Error> {
+        Ok(self.is_low())
+    }
+}
+
+impl<MODE, const P: char> PXx<Input<MODE>, P> {
+    #[inline(always)]
+    fn is_high(&self) -> bool {
+        !self.is_low()
+    }
+
+    #[inline(always)]
+    fn is_low(&self) -> bool {
         // NOTE(unsafe) atomic read with no side effects
-        Ok(unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << self.i) == 0 })
+        unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << self.i) == 0 }
     }
 }
 
 impl<MODE, const P: char> InputPin for PXx<Input<MODE>, P> {
     type Error = Infallible;
 
+    #[inline(always)]
     fn is_high(&self) -> Result<bool, Self::Error> {
-        self.is_low().map(|v| !v)
+        Ok(self.is_high())
     }
 
+    #[inline(always)]
     fn is_low(&self) -> Result<bool, Self::Error> {
-        // NOTE(unsafe) atomic read with no side effects
-        Ok(unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << self.i) == 0 })
+        Ok(self.is_low())
     }
 }
 
@@ -326,17 +424,337 @@ impl<MODE, const P: char, const N: u8> PinExt for PX<MODE, P, N> {
     }
 }
 
+impl<MODE, const P: char, const N: u8, const A: u8> From<PX<Input<MODE>, P, N>>
+    for PX<Alternate<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Input<MODE>, P, N>) -> Self {
+        f.into_alternate::<A>()
+    }
+}
+
+impl<MODE, const P: char, const N: u8, const A: u8> From<PX<Output<MODE>, P, N>>
+    for PX<Alternate<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Output<MODE>, P, N>) -> Self {
+        f.into_alternate::<A>()
+    }
+}
+
+impl<const P: char, const N: u8, const A: u8> From<PX<Analog, P, N>>
+    for PX<Alternate<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Analog, P, N>) -> Self {
+        f.into_alternate::<A>()
+    }
+}
+
+impl<const P: char, const N: u8, const A: u8, const B: u8> From<PX<AlternateOD<AF<B>>, P, N>>
+    for PX<Alternate<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<AlternateOD<AF<B>>, P, N>) -> Self {
+        f.into_alternate::<A>()
+    }
+}
+
+impl<MODE, const P: char, const N: u8, const A: u8> From<PX<Input<MODE>, P, N>>
+    for PX<AlternateOD<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Input<MODE>, P, N>) -> Self {
+        f.into_alternate_open_drain::<A>()
+    }
+}
+
+impl<MODE, const P: char, const N: u8, const A: u8> From<PX<Output<MODE>, P, N>>
+    for PX<AlternateOD<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Output<MODE>, P, N>) -> Self {
+        f.into_alternate_open_drain::<A>()
+    }
+}
+
+impl<const P: char, const N: u8, const A: u8> From<PX<Analog, P, N>>
+    for PX<AlternateOD<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Analog, P, N>) -> Self {
+        f.into_alternate_open_drain::<A>()
+    }
+}
+
+impl<const P: char, const N: u8, const A: u8, const B: u8> From<PX<Alternate<AF<B>>, P, N>>
+    for PX<AlternateOD<AF<A>>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Alternate<AF<B>>, P, N>) -> Self {
+        f.into_alternate_open_drain::<A>()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Input<Floating>, P, N>> for PX<Input<PullDown>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<Floating>, P, N>) -> Self {
+        f.into_pull_down_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Input<PullUp>, P, N>> for PX<Input<PullDown>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<PullUp>, P, N>) -> Self {
+        f.into_pull_down_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Output<MODE>, P, N>> for PX<Input<PullDown>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Output<MODE>, P, N>) -> Self {
+        f.into_pull_down_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Analog, P, N>> for PX<Input<PullDown>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Analog, P, N>) -> Self {
+        f.into_pull_down_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Alternate<MODE>, P, N>>
+    for PX<Input<PullDown>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Alternate<MODE>, P, N>) -> Self {
+        f.into_pull_down_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<AlternateOD<MODE>, P, N>>
+    for PX<Input<PullDown>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<AlternateOD<MODE>, P, N>) -> Self {
+        f.into_pull_down_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Input<Floating>, P, N>> for PX<Input<PullUp>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<Floating>, P, N>) -> Self {
+        f.into_pull_up_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Input<PullDown>, P, N>> for PX<Input<PullUp>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<PullDown>, P, N>) -> Self {
+        f.into_pull_up_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Output<MODE>, P, N>> for PX<Input<PullUp>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Output<MODE>, P, N>) -> Self {
+        f.into_pull_up_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Analog, P, N>> for PX<Input<PullUp>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Analog, P, N>) -> Self {
+        f.into_pull_up_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Alternate<MODE>, P, N>> for PX<Input<PullUp>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Alternate<MODE>, P, N>) -> Self {
+        f.into_pull_up_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<AlternateOD<MODE>, P, N>>
+    for PX<Input<PullUp>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<AlternateOD<MODE>, P, N>) -> Self {
+        f.into_pull_up_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Input<PullDown>, P, N>> for PX<Input<Floating>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<PullDown>, P, N>) -> Self {
+        f.into_floating_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Input<PullUp>, P, N>> for PX<Input<Floating>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<PullUp>, P, N>) -> Self {
+        f.into_floating_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Output<MODE>, P, N>> for PX<Input<Floating>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Output<MODE>, P, N>) -> Self {
+        f.into_floating_input()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Analog, P, N>> for PX<Input<Floating>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Analog, P, N>) -> Self {
+        f.into_floating_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Alternate<MODE>, P, N>>
+    for PX<Input<Floating>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Alternate<MODE>, P, N>) -> Self {
+        f.into_floating_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<AlternateOD<MODE>, P, N>>
+    for PX<Input<Floating>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<AlternateOD<MODE>, P, N>) -> Self {
+        f.into_floating_input()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Input<MODE>, P, N>> for PX<Output<OpenDrain>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<MODE>, P, N>) -> Self {
+        f.into_open_drain_output()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Output<PushPull>, P, N>> for PX<Output<OpenDrain>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Output<PushPull>, P, N>) -> Self {
+        f.into_open_drain_output()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Analog, P, N>> for PX<Output<OpenDrain>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Analog, P, N>) -> Self {
+        f.into_open_drain_output()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Alternate<MODE>, P, N>>
+    for PX<Output<OpenDrain>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Alternate<MODE>, P, N>) -> Self {
+        f.into_open_drain_output()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<AlternateOD<MODE>, P, N>>
+    for PX<Output<OpenDrain>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<AlternateOD<MODE>, P, N>) -> Self {
+        f.into_open_drain_output()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Input<MODE>, P, N>> for PX<Output<PushPull>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<MODE>, P, N>) -> Self {
+        f.into_push_pull_output()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Output<OpenDrain>, P, N>> for PX<Output<PushPull>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Output<OpenDrain>, P, N>) -> Self {
+        f.into_push_pull_output()
+    }
+}
+
+impl<const P: char, const N: u8> From<PX<Analog, P, N>> for PX<Output<PushPull>, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Analog, P, N>) -> Self {
+        f.into_push_pull_output()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Alternate<MODE>, P, N>>
+    for PX<Output<PushPull>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<Alternate<MODE>, P, N>) -> Self {
+        f.into_push_pull_output()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<AlternateOD<MODE>, P, N>>
+    for PX<Output<PushPull>, P, N>
+{
+    #[inline(always)]
+    fn from(f: PX<AlternateOD<MODE>, P, N>) -> Self {
+        f.into_push_pull_output()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Input<MODE>, P, N>> for PX<Analog, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Input<MODE>, P, N>) -> Self {
+        f.into_analog()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Output<MODE>, P, N>> for PX<Analog, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Output<MODE>, P, N>) -> Self {
+        f.into_analog()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<Alternate<MODE>, P, N>> for PX<Analog, P, N> {
+    #[inline(always)]
+    fn from(f: PX<Alternate<MODE>, P, N>) -> Self {
+        f.into_analog()
+    }
+}
+
+impl<MODE, const P: char, const N: u8> From<PX<AlternateOD<MODE>, P, N>> for PX<Analog, P, N> {
+    #[inline(always)]
+    fn from(f: PX<AlternateOD<MODE>, P, N>) -> Self {
+        f.into_analog()
+    }
+}
 impl<MODE, const P: char, const N: u8> PX<MODE, P, N> {
     /// Configures the pin to operate alternate mode
     pub fn into_alternate<const A: u8>(self) -> PX<Alternate<AF<A>>, P, N> {
-        less_than_16::<A>();
+        #[allow(path_statements, clippy::no_effect)]
+        {
+            Assert::<A, 16>::LESS;
+        }
         _set_alternate_mode::<P, N, A>();
         PX::new()
     }
 
     /// Configures the pin to operate in alternate open drain mode
+    #[allow(path_statements)]
     pub fn into_alternate_open_drain<const A: u8>(self) -> PX<AlternateOD<AF<A>>, P, N> {
-        less_than_16::<A>();
+        #[allow(path_statements, clippy::no_effect)]
+        {
+            Assert::<A, 16>::LESS;
+        }
         _set_alternate_mode::<P, N, A>();
         PX::new().set_open_drain()
     }
@@ -754,62 +1172,150 @@ impl<MODE, const P: char, const N: u8> PX<MODE, P, N> {
     }
 }
 
-impl<MODE, const P: char, const N: u8> OutputPin for PX<Output<MODE>, P, N> {
-    type Error = Infallible;
-
-    fn set_high(&mut self) -> Result<(), Self::Error> {
+impl<MODE, const P: char, const N: u8> PX<Output<MODE>, P, N> {
+    #[inline(always)]
+    pub fn set_high(&mut self) {
         // NOTE(unsafe) atomic write to a stateless register
-        unsafe { (*Gpio::<P>::ptr()).bsrr.write(|w| w.bits(1 << { N })) };
-        Ok(())
+        unsafe { (*Gpio::<P>::ptr()).bsrr.write(|w| w.bits(1 << { N })) }
     }
 
-    fn set_low(&mut self) -> Result<(), Self::Error> {
+    #[inline(always)]
+    pub fn set_low(&mut self) {
         // NOTE(unsafe) atomic write to a stateless register
         unsafe {
             (*Gpio::<P>::ptr())
                 .bsrr
                 .write(|w| w.bits(1 << ({ N } + 16)))
-        };
+        }
+    }
+
+    #[inline(always)]
+    pub fn get_state(&self) -> PinState {
+        if self.is_set_low() {
+            PinState::Low
+        } else {
+            PinState::High
+        }
+    }
+
+    #[inline(always)]
+    pub fn set_state(&mut self, state: PinState) {
+        match state {
+            PinState::Low => self.set_low(),
+            PinState::High => self.set_high(),
+        }
+    }
+
+    #[inline(always)]
+    pub fn is_set_high(&self) -> bool {
+        !self.is_set_low()
+    }
+
+    #[inline(always)]
+    pub fn is_set_low(&self) -> bool {
+        // NOTE(unsafe) atomic read with no side effects
+        unsafe { (*Gpio::<P>::ptr()).odr.read().bits() & (1 << { N }) == 0 }
+    }
+
+    #[inline(always)]
+    pub fn toggle(&mut self) {
+        if self.is_set_low() {
+            self.set_high()
+        } else {
+            self.set_low()
+        }
+    }
+}
+
+impl<MODE, const P: char, const N: u8> OutputPin for PX<Output<MODE>, P, N> {
+    type Error = Infallible;
+
+    #[inline(always)]
+    fn set_high(&mut self) -> Result<(), Self::Error> {
+        self.set_high();
+        Ok(())
+    }
+
+    #[inline(always)]
+    fn set_low(&mut self) -> Result<(), Self::Error> {
+        self.set_low();
         Ok(())
     }
 }
 
 impl<MODE, const P: char, const N: u8> StatefulOutputPin for PX<Output<MODE>, P, N> {
+    #[inline(always)]
     fn is_set_high(&self) -> Result<bool, Self::Error> {
-        self.is_set_low().map(|v| !v)
+        Ok(self.is_set_high())
     }
 
+    #[inline(always)]
     fn is_set_low(&self) -> Result<bool, Self::Error> {
-        // NOTE(unsafe) atomic read with no side effects
-        Ok(unsafe { (*Gpio::<P>::ptr()).odr.read().bits() & (1 << { N }) == 0 })
+        Ok(self.is_set_low())
     }
 }
 
-impl<MODE, const P: char, const N: u8> toggleable::Default for PX<Output<MODE>, P, N> {}
+impl<MODE, const P: char, const N: u8> ToggleableOutputPin for PX<Output<MODE>, P, N> {
+    type Error = Infallible;
+
+    #[inline(always)]
+    fn toggle(&mut self) -> Result<(), Self::Error> {
+        self.toggle();
+        Ok(())
+    }
+}
+
+impl<const P: char, const N: u8> PX<Output<OpenDrain>, P, N> {
+    #[inline(always)]
+    pub fn is_high(&self) -> bool {
+        !self.is_low()
+    }
+
+    #[inline(always)]
+    pub fn is_low(&self) -> bool {
+        // NOTE(unsafe) atomic read with no side effects
+        unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << { N }) == 0 }
+    }
+}
 
 impl<const P: char, const N: u8> InputPin for PX<Output<OpenDrain>, P, N> {
     type Error = Infallible;
 
+    #[inline(always)]
     fn is_high(&self) -> Result<bool, Self::Error> {
-        self.is_low().map(|v| !v)
+        Ok(self.is_high())
     }
 
+    #[inline(always)]
     fn is_low(&self) -> Result<bool, Self::Error> {
+        Ok(self.is_low())
+    }
+}
+
+impl<MODE, const P: char, const N: u8> PX<Input<MODE>, P, N> {
+    #[inline(always)]
+    pub fn is_high(&self) -> bool {
+        !self.is_low()
+    }
+
+    #[inline(always)]
+    pub fn is_low(&self) -> bool {
         // NOTE(unsafe) atomic read with no side effects
-        Ok(unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << { N }) == 0 })
+        unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << { N }) == 0 }
     }
 }
 
 impl<MODE, const P: char, const N: u8> InputPin for PX<Input<MODE>, P, N> {
     type Error = Infallible;
 
+    #[inline(always)]
     fn is_high(&self) -> Result<bool, Self::Error> {
-        self.is_low().map(|v| !v)
+        Ok(self.is_high())
     }
 
+    #[inline(always)]
     fn is_low(&self) -> Result<bool, Self::Error> {
-        // NOTE(unsafe) atomic read with no side effects
-        Ok(unsafe { (*Gpio::<P>::ptr()).idr.read().bits() & (1 << { N }) == 0 })
+        Ok(self.is_low())
     }
 }
 
@@ -1126,60 +1632,147 @@ impl<MODE> Pin<MODE> {
     }
 }
 
-impl<MODE> OutputPin for Pin<Output<MODE>> {
-    type Error = core::convert::Infallible;
+impl<MODE> Pin<Output<MODE>> {
+    #[inline(always)]
+    pub fn set_high(&mut self) {
+        // NOTE(unsafe) atomic write to a stateless register
+        unsafe { self.block().bsrr.write(|w| w.bits(1 << self.pin_id())) };
+    }
 
-    fn set_low(&mut self) -> Result<(), Self::Error> {
+    #[inline(always)]
+    pub fn set_low(&mut self) {
         // NOTE(unsafe) atomic write to a stateless register
         unsafe {
             self.block()
                 .bsrr
                 .write(|w| w.bits(1 << (self.pin_id() + 16)))
         };
+    }
+
+    #[inline(always)]
+    pub fn get_state(&self) -> PinState {
+        if self.is_set_low() {
+            PinState::Low
+        } else {
+            PinState::High
+        }
+    }
+
+    #[inline(always)]
+    pub fn set_state(&mut self, state: PinState) {
+        match state {
+            PinState::Low => self.set_low(),
+            PinState::High => self.set_high(),
+        }
+    }
+
+    #[inline(always)]
+    pub fn is_set_high(&self) -> bool {
+        !self.is_set_low()
+    }
+
+    #[inline(always)]
+    pub fn is_set_low(&self) -> bool {
+        self.block().odr.read().bits() & (1 << self.pin_id()) == 0
+    }
+
+    #[inline(always)]
+    pub fn toggle(&mut self) {
+        if self.is_set_low() {
+            self.set_high()
+        } else {
+            self.set_low()
+        }
+    }
+}
+
+impl<MODE> OutputPin for Pin<Output<MODE>> {
+    type Error = core::convert::Infallible;
+
+    #[inline(always)]
+    fn set_high(&mut self) -> Result<(), Self::Error> {
+        self.set_high();
         Ok(())
     }
 
-    fn set_high(&mut self) -> Result<(), Self::Error> {
-        // NOTE(unsafe) atomic write to a stateless register
-        unsafe { self.block().bsrr.write(|w| w.bits(1 << self.pin_id())) };
+    #[inline(always)]
+    fn set_low(&mut self) -> Result<(), Self::Error> {
+        self.set_low();
         Ok(())
     }
 }
 
 impl<MODE> StatefulOutputPin for Pin<Output<MODE>> {
+    #[inline(always)]
     fn is_set_high(&self) -> Result<bool, Self::Error> {
-        self.is_set_low().map(|v| !v)
+        Ok(self.is_set_high())
     }
 
+    #[inline(always)]
     fn is_set_low(&self) -> Result<bool, Self::Error> {
-        // NOTE(unsafe) atomic read with no side effects
-        Ok(self.block().odr.read().bits() & (1 << self.pin_id()) == 0)
+        Ok(self.is_set_low())
     }
 }
 
-impl<MODE> toggleable::Default for Pin<Output<MODE>> {}
+impl<MODE> ToggleableOutputPin for Pin<Output<MODE>> {
+    type Error = Infallible;
+
+    #[inline(always)]
+    fn toggle(&mut self) -> Result<(), Self::Error> {
+        self.toggle();
+        Ok(())
+    }
+}
+
+impl Pin<Output<OpenDrain>> {
+    #[inline(always)]
+    fn is_high(&self) -> bool {
+        !self.is_low()
+    }
+
+    #[inline(always)]
+    fn is_low(&self) -> bool {
+        self.block().idr.read().bits() & (1 << self.pin_id()) == 0
+    }
+}
 
 impl InputPin for Pin<Output<OpenDrain>> {
     type Error = core::convert::Infallible;
 
+    #[inline(always)]
     fn is_high(&self) -> Result<bool, Self::Error> {
-        self.is_low().map(|v| !v)
+        Ok(self.is_high())
     }
 
+    #[inline(always)]
     fn is_low(&self) -> Result<bool, Self::Error> {
-        Ok(self.block().idr.read().bits() & (1 << self.pin_id()) == 0)
+        Ok(self.is_low())
+    }
+}
+
+impl<MODE> Pin<Input<MODE>> {
+    #[inline(always)]
+    fn is_high(&self) -> bool {
+        !self.is_low()
+    }
+
+    #[inline(always)]
+    fn is_low(&self) -> bool {
+        self.block().idr.read().bits() & (1 << self.pin_id()) == 0
     }
 }
 
 impl<MODE> InputPin for Pin<Input<MODE>> {
     type Error = core::convert::Infallible;
 
+    #[inline(always)]
     fn is_high(&self) -> Result<bool, Self::Error> {
-        self.is_low().map(|v| !v)
+        Ok(self.is_high())
     }
 
+    #[inline(always)]
     fn is_low(&self) -> Result<bool, Self::Error> {
-        Ok(self.block().idr.read().bits() & (1 << self.pin_id()) == 0)
+        Ok(self.is_low())
     }
 }
 
@@ -1208,11 +1801,6 @@ impl<const P: char> Gpio<P> {
             _ => crate::pac::GPIOA::ptr(),
         }
     }
-}
-
-#[allow(path_statements)]
-const fn less_than_16<const A: u8>() {
-    Assert::<A, 16>::LESS;
 }
 
 /// Const assert hack